+++ public/stylesheets/style.css
@@ -34,6 +34,10 @@
     color: white;
 }
 
+a.navbar-item:hover {
+    color: aquamarine;
+}
+
 #profile, #logout {
     color: black;
     text-decoration: none;
