+++ views/homepage.handlebars
@@ -2,10 +2,10 @@
 <section class="main-pg section is-medium">
     <div class="columns is-centered">
         <div class="column is-5">
-            <h1 class="center is-size-3">Get Inspired! Click the button to generate a random chord progression.</h1>
+            <h1 class="center is-size-3 has-text-success">Get Inspired! Click the button to generate a random chord progression.</h1>
              <div id="display" class="hidden mx-6">
-                <div class="box mx-6">
-                    <div id='chord-columns' class="columns is-mobile is-centered">
+                <div class="box">
+                    <div id='chord-columns' class="columns is-centered">
                         {{!-- chords with be dynamically generated here --}}
                     </div>
                 </div>
@@ -17,7 +17,7 @@
     </div>
     
     <div class="login columns is-centered">
-        <p class="column is-5 center">or <a href="/login">login/signup</a> for more music inspiration!</p>
+        <p class="column is-5 center">or <a href="/login" class="has-text-success">login/signup</a> for more music inspiration!</p>
     </div>
 
 </section>
