+++ views/login.handlebars
@@ -20,7 +20,7 @@
             </form>
             <div>
                 <div class="center">
-                    or <a href="/signup">signup</a>
+                    or <a class="has-text-success" href="/signup">signup</a>
                 </div>
             </div>
         </div>
